fixup handling of existing branches on remote
list remote for matching branches and remove them
if they exist. This is destructive and perhaps undesirable
but the preference here is to do the same thing every time, e.g
  cut a branch
  add & commit the formatted files
  push that branch
  open a pull request

Acknowledging that it wouldn't be difficult to simply push to a branch
that's already existing on remote, the intention here is to do the same
thing every time.